requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
pytest>=6.0.0
pytest-cov>=2.10.0
faiss-cpu>=1.7.0
pydantic>=1.8.0
PyYAML>=5.4.0
pydantic>=1.8.0
PyYAML>=5.4.0